+++ src/index.js
@@ -15,6 +15,7 @@ import LibrarySong from 'components/library/song/List'
 import LibraryWork from 'components/library/work/List'
 import Main from 'components/Main'
 import NotFound from 'components/navigation/NotFound'
+import PlayerErrors from 'components/playlist/playerErrors/List'
 import PlaylistPlayed from 'components/playlist/played/List'
 import Playlist from 'components/playlist/Playlist'
 import PlaylistQueueing from 'components/playlist/queueing/List'
@@ -119,6 +120,10 @@ root.render(
                             <Route
                                 path="played"
                                 element={(<PlaylistPlayed />)}
+                            />
+                            <Route
+                                path="player-errors"
+                                element={(<PlayerErrors/>)}
                             />
                         </Route>
                         <Route path="settings" element={(<Settings />)}>