+++ CHANGELOG.md
@@ -30,6 +30,10 @@
 
 ## Unreleased
 
+### Added
+
+- Player errors are now displayed in a library fashion.
+
 ### Changed
 
 - Played and queuing playlist entries can be accessed in a library fashion: they use pagination now, and queuing entries can be send to the first or the last place.